Move paper formats to own module
